% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_inv.R
\name{mortality_model}
\alias{mortality_model}
\title{Mortality Prediction}
\usage{
mortality_model(input)
}
\arguments{
\item{input}{Mortality data}
}
\value{
Predicted mortality rates
}
\description{
Fit poisson GLM to predict mortality rates
}
\examples{
predictions <- mortality_model("mortality_data.csv");
}
